requirements.txt
python-telegram-bot==20.7
beautifulsoup4==4.12.2
requests==2.31.0
python-dotenv==1.0.0
lxml==4.9.3
reportlab==4.0.7
weasyprint==60.2
psycopg2-binary==2.9.9
